Fix GH#97 and GH#115: enforce external crate ownership boundaries
GH#97: Move hook SQL out of hooks crate into db
- Move Hook domain types (Hook, HookSource, HookAction, HookExecution, etc.)
  from hooks to types crate
- Move HookStore trait + SqliteHookStore impl from hooks::store into db crate,
  reusing db::Error for store errors
- Remove sqlx direct dep from hooks/Cargo.toml and server/Cargo.toml
- server creates SqliteHookStore via db.pool() instead of a separate sqlx pool
- routes/hook.rs uses db::Error instead of hooks::store::StoreError

GH#115: Add external crate ownership arch tests (one per constraint)
- only_db_uses_sqlx_directly
- only_server_uses_axum_directly
- only_designated_http_crates_use_reqwest_directly

New ns2-client crate: The new reqwest arch test revealed cli also owned reqwest
directly (for local server HTTP calls). Extracted into a new ns2-client crate
that owns reqwest for ns2 server communication, analogous to how anthropic
owns reqwest for the Anthropic API. cli now depends on ns2-client.
Renamed cli's HookAction subcommand enum to HookSubcommand to avoid collision
with types::HookAction.
