+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "crates/server",
     "crates/tui",
     "crates/cli",
+    "crates/ns2-client",
     "crates/arch-tests",
     "crates/specs",
 ]
