+++ crates/cli/Cargo.toml
@@ -13,7 +13,7 @@ thiserror = { workspace = true }
 tokio = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-reqwest = { version = "0.12", features = ["json", "stream"] }
+ns2-client = { path = "../ns2-client" }
 types = { path = "../types" }
 events = { path = "../events" }
 workspace = { path = "../workspace" }
