+++ crates/hooks/Cargo.toml
@@ -13,11 +13,8 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 uuid = { workspace = true }
 chrono = { workspace = true }
-thiserror = { workspace = true }
-async-trait = { workspace = true }
 tracing = { workspace = true }
 minijinja = "2"
-sqlx = { version = "0.7", features = ["sqlite", "runtime-tokio-native-tls"] }
 
 [dev-dependencies]
 tokio = { workspace = true }
